Merge branch '多层' into 'main': 支持多层

+++ utils/fileUtils/fileUtls.go
@@ -2,6 +2,8 @@ package fileUtils
 
 import (
 	"bufio"
+	"dst-admin-go/utils/systemUtils"
+	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -38,7 +40,7 @@ func CreateDir(dirName string) bool {
 	if Exists(dirName) {
 		return false
 	}
-	err := os.Mkdir(dirName, 0755)
+	err := os.MkdirAll(dirName, 0755)
 	if err != nil {
 		log.Println(err)
 		return false
@@ -144,29 +146,42 @@ func WriterLnFile(filename string, lines []string) error {
 	return w.Flush()
 }
 
-func ReverseRead(filename string, n uint) ([]string, error) {
+func ReverseRead(filename string, lineNum uint) ([]string, error) {
 	file, err := os.Open(filename)
 	if err != nil {
 		return nil, err
 	}
 	defer file.Close()
+	//获取文件大小
+	fs, err := file.Stat()
+	fileSize := fs.Size()
 
-	lines := make([]string, 0, n)
-	scanner := bufio.NewScanner(file)
-
-	// Read lines in reverse order
-	for scanner.Scan() {
-		lines = append([]string{scanner.Text()}, lines...)
-		if len(lines) > int(n) {
-			lines = lines[:n]
+	var offset int64 = -1   //偏移量，初始化为-1，若为0则会读到EOF
+	char := make([]byte, 1) //用于读取单个字节
+	lineStr := ""           //存放一行的数据
+	buff := make([]string, 0, 100)
+	for (-offset) <= fileSize {
+		//通过Seek函数从末尾移动游标然后每次读取一个字节
+		file.Seek(offset, io.SeekEnd)
+		_, err := file.Read(char)
+		if err != nil {
+			return buff, err
 		}
+		if char[0] == '\n' {
+			// offset--  //windows跳过'\r'
+			lineNum-- //到此读取完一行
+			buff = append(buff, lineStr)
+			lineStr = ""
+			if lineNum == 0 {
+				return buff, nil
+			}
+		} else {
+			lineStr = string(char) + lineStr
+		}
+		offset--
 	}
-
-	if err := scanner.Err(); err != nil {
-		return nil, err
-	}
-
-	return lines, nil
+	buff = append(buff, lineStr)
+	return buff, nil
 }
 
 func DeleteFile(path string) error {
@@ -178,12 +193,19 @@ func DeleteFile(path string) error {
 	return nil
 }
 
-func DeleteDir(path string) (err error) {
+func DeleteDir(path string) error {
+	home, err := systemUtils.Home()
+	if err != nil {
+		return err
+	}
+	if path == filepath.Join(home, ".klei/DoNotStarveTogether") {
+		return errors.New(".klei/DoNotStarveTogether not allow to delete")
+	}
 	err = os.RemoveAll(path)
 	if err != nil {
-		log.Printf("removeAll "+path+" : %v\n", err)
+		log.Printf("remove err "+path+" : %v\n", err)
 	}
-	return
+	return err
 }
 
 func Rename(filePath, newName string) (err error) {
